Coerce Python 2 objects via old buffer interface
As Python 2 objects can be comfortably coerced to the old-style buffer
interface, which can cleanly be converted to a `memoryview`, go ahead
and coerce everything to an old-style buffer in Python 2. Then it is a
straightforward matter to get a `memoryview` from the object and handle
it the same as one might on Python 3.