Smooth over Python 2/3 memoryview coercion
In Python 2, everything can be comfortably coerced to the old-style
buffer interface, which is easily coerced to the new buffer interface.
This is a nice path to proceed down as it can cleanly get a `memoryview`
on Python 2 without copying. This strategy works well on Python 3 as
well as long as we make `buffer` a no-op.