Coerce data to old-style buffers on Python 2
In the GZip and Zlib encoders, coerce data to old-style buffers on
Python 2 as both GZip and Zlib can handle these inputs. By doing this,
we are able to avoid inducing a copy as would have been the case when
converting the data to `bytes`. We also are able to handle a wide
variety of inputs include `bytearray`s and `array`s.