Alias `buffer` as `memoryview` on Python 3
This is sort of a lie. That said, they are pretty close to one another
and most compat documentation suggests to do this to migrate Python 2
code to Python 3. We only need this so importing `buffer` to appease
`flake8` doesn't become even more of a mess, which seems a bit odd.