Add alias to `buffer` on Python 2/3
While `buffer` is not exactly equivalent to `memoryview` on Python 3 and
Python 2 already has `memoryview`, the support of the `buffer` type is
much better within Python 2 much as the support of `memoryview` is much
better on Python 3. Plus this is what a lot of Python 2/3 compatibility
guides suggest when moving to Python 3. So this seems like a reasonable
alias to add.