Use `to_buffer` to implement `buffer_tobytes`
Effectively `to_buffer` is a refactoring of the functionality in
`buffer_tobytes`. Namely it handles taking a view of the underlying data
and ensuring that view is flattened, contiguous, and cast to `unsigned
char`. Though it doesn't actually perform the copy yet.

Hence `buffer_tobytes` leverages `to_buffer` to ensure the data is ready
to copy. All that is needed is coercion to a `memoryview`. Once coerced
the `tobytes` method can be called to return the `bytes` needed for the
encoding or decoding operation. This causes a copy, but is unavoidable
at this stage.