Add `to_buffer` function
Provides a compatibility function for converting an object into a buffer
or memoryview. Makes sure the Python 2 and Python 3 buffers behave
roughly the same. Use the `buffer` type on Python 2 as more things
support it and it can easily be coerced to a `memoryview`. On Python 3
use a `memoryview` as there is no `buffer` type and `memoryview`s work
well on Python 3. Ensure the Python 3 `memoryview` is flattened and of
`unsigned char` type, which is inline with the `memoryview` coerced
Python 2 buffer. More importantly we generally want to work with
flattened byte types anyways, so this provides us a consistent path to
getting there while still providing a view onto the underlying data
(without copying on either Python).