Use `to_buffer` in `GZip` encoding and decoding
Ensures that we are working with a consistent view onto the underlying
data with both Python 2 and Python 3. The `buffer` or `memoryview`
returned on Python 2 or Python 3 respectively can be comfortably used
with `BytesIO` and `GzipFile`'s `write` with the minimal necessary
overhead. Also removes the need for functions like `handle_datetime`, as
`to_buffer` already ensures an unsigned character view of the data.