Use `to_buffer` in `Zlib` encoding and decoding
Ensures that we are working with a consistent view onto the underlying
data with both Python 2 and Python 3. The `buffer` or `memoryview`
returned on Python 2 or Python 3 respectively can be comfortably used
with `zlib`'s `compress` and `decompress` functions with the minimal
necessary overhead. Also drops the need for `handle_datetime` as
`to_buffer` already takes an unsigned character view of the data.