Raise in `to_buffer` if `object` array is provided
As encoding `object` arrays to `bytes` does not really make sense (one
gets address of pointers on the stack), raise if any data provided is of
`object` type. This can be done easily with an `ndarray`. In other
cases, it can be checked through a `memoryview` of the data. Though
there are some edge cases on Python 2, which need special handling.