Use `buffer_tobytes` on non-`object` types in test
When running the `check_backwards_compatibility` test, make sure not to
call `buffer_tobytes` on `object` arrays from fixture data. Casting
these to bytes doesn't really make sense as it will return
representation of pointers in the underlying buffer. Not only are the
pointers things we do not want to be comparing, but it is also internal
spec to NumPy and Python that really isn't reliable either. So make sure
to pass through the `object` arrays as is.