Simply catch all non-`memoryview` coercible types
When attempt to coerce an object to a `memoryview` to determine the
buffer's type on Python 2, just use a `try...catch` to ignore all cases
that fail. The are generally things like `array` or `mmap`, which cannot
hold object types anyways. So they can be presumed safe and ignored.

We are only interested in checking for things like `memoryviews` of
other things that already contain `object`s or some third-party
`ndarray`-like class that holds `object`s. If any of these show up,
throw a `ValueError` per usual.