Breakup array handling, object check, & Python 2/3
To simplify the code a bit, split it up into chunks. First handle the
flattening of arrays and coercion of datetime/timedelta types. Second
check to see if the data is actually `object` type and raise if it is.
Third handle Python 2/3 differences of `buffer`s and `memoryview`s
respectively. This cuts down the code and makes the remaining code
easier to read and follow. Also should make it less prone to edge case
bugs.