Reshape all `memoryview`s using NumPy
Instead of simply reshaping NumPy arrays, take anything that can be
coerced into a `memoryview` and view it as a NumPy array so as to
reshape it. This way all `memoryview`s (including NumPy arrays) can be
consistently reshaped. Not to mention this works on Python 2 and Python
3.

By doing this, we are able to drop usage of `memoryview`'s `cast`
method. This was restricted to Python 3. So provided no options for
handling equivalent cases in Python 2. Also the `cast` method was a bit
convoluted for just reshaping. Not to mention it was unclear that this
would enforce C or Fortran ordering (though it did seem to raise if
those were not achievable).

Of course this doesn't work for Python 2 `buffer` objects, but we
weren't doing anything special for them anyways. Besides the few cases
where Python 2 cannot construct a `memoryview` are for things like
`array` and `mmap`, which are contiguous and flat anyways. Given how the
old `buffer` object could handle chunks of non-contiguous memory,
reshaping or forcing contiguous memory seems unnecessary in that case.