Mark `to_buffer`'s NumPy array as readonly
As the point of using `to_buffer` is to coerce external data into a
buffer that matches are constraints and can be used efficiently to feed
data into encoders/decoders, the buffer should not ever need to be
writable (only readable). Given this, mark the flattened `ndarray` of
our data as readonly. By doing this, we ensure that `memoryview`s
constructed from this `ndarray` will also be `readonly`. Thus protecting
the user's data from accidentally being overwritten.