Test `to_buffer` returns readonly buffers
Includes a simple test of passing in writable array data to `to_buffer`
and checks that the buffer returned has its `readonly` attribute
enabled. On Python 2 this is always true as `buffer` creates readonly
buffers only, but it is not easy to inspect. So we coerce the `buffer`
object to a `memoryview`, which has the `readonly` attribute. On Python
3, this extra coercion is cheap since we have a `memoryview` to begin
with. So it should just copy the existing `memoryview` on Python 3
(though not the underlying data).