Coerce `memoryview`s and `buffer`s to `ndarray`s
Instead of returning a `memoryview` or `buffer` from `to_buffer`, return
something coercible to either type. Namely return an `ndarray` that
views, but does not copy, the underlying data. That way compression
algorithms can easily coerce the data into a `memoryview`, `buffer`, or
`bytes` depending on which is appropriate and works best for the data.