Ensure `array` has the correct type on Python 2
On Python 2, the `array` type cannot be cast to a `memoryview`. Though
it can be cast to a `buffer`. However the old `buffer` interface does
not keep track of the type of data it has and merely represents it as
`unsigned char`. This behavior is reproduced by `ndarray`s viewing the
`buffer` as well. To fix this, we cast the `ndarray` back to the type
specified by the `array`.