Limit casting of `array` to Python 2
On Python 3, `array` is coercible to a `memoryview`. So it maintains its
type information and the `ndarray` will reflect that. While casting to
the `ndarray` to the type it already is won't hurt, it isn't needed
either. So restrict this to Python 2, where `array` can only be coerced
to `buffer` and the type information must be restored.