Handle reshaping outside of `ndarray_from_buffer`
Performing reshaping within `ndarray_from_buffer` makes this general
function a little harder to use. Especially as the reshaping may cause
an unwanted copy. Given this is a trivial thing for users to do
externally, drop the reshape from the `ndarray_from_buffer` and rewrite
usages of `ndarray_from_buffer` to handle the reshaping themselves.