Raise for unicode `array` in `ndarray_from_buffer`
Handling of unicode `array`s are generally a bit broken. NumPy doesn't
have native support of the type. Trying to workaround their issues in a
cross-platform Python 2/3 friendly way is fairly difficult. Also they
have been deprecated since Python 3.3. Plus these are likely an
exceedingly rare thing to see in use. Given all of these issues, simply
raise an error if users provide a unicode array stating they are not
supported. If it turns out there are users desperately in need of this
feature, we can always revisit and discus with them. However we will
really need their feedback as any implementation here is going to
require feedback and likely someone dedicated to solving the problem.
For now erroring out is the most simplest path forward.

ref: https://bugs.python.org/issue13072
ref: https://bugs.python.org/issue15625