Cast around unicode array
On Python 3, cast the unicode array to an unsigned char array. This
matches what will happen on Python 2 as well. Should bypass NumPy and
Windows issues on Python 3 with this type as well because the casting
precedes construction of the array. As Python 2 already goes through
this path, this is a non-issue there. Afterwards cast the resulting
`ndarray` to an `unsigned` type with the same `itemsize` as the original
unicode type (as it may be 16-bit or 32-bit depending on the Python
build). Handle all other `array` cases the same as before.