Preserve `memoryview` of `array`'s `itemsize`
Perform both `memoryview` casts for unicode `array`s one after the
other. This ensures that the `memoryview` has the expected replacement
type before taking the `ndarray` view. The behavior for an end user
appears to be essentially identical. However it's arguably nicer to
ensure the underlying `memoryview` has the intended type and shape.
Updates the `ndarray` casting block at the end to be Python 2 only. As
the unicode `array` case (the only one in need of casting on Python 3),
is handled before constructing the `ndarray`. So there is no need for
additional casting of `ndarray`s for Python 3.