Cast `memoryview` to `uint8` for `shares_memory`
As the buffer is coerced by NumPy into an array without copying, it
needs to be able to handle the data type. However for Windows the
unicode array has UCS2 whereas NumPy only supports UCS4. To fix this
issue just cast the `memoryview` to `uint8`. We don't particularly care
about the type for this test. Only that data is shared. So this should
workaround the issue and allow the check to occur.