Ensure `encode` returns a `ndarray`
Standardize the output of the `encode` function a bit by ensuring it
returns an `ndarray`. This is done without copying as it only takes a
view onto the data. Has the advantage of leveraging objects under the
hood that might not appear as nice like objects' internal buffers
without impacting the user or unnecessarily copying the data to `bytes`.
Also makes it a bit easier to work with the output as users can do
anything they would normally want to with `ndarray`s.