Return `ndarray`s from `decode` where possible
Where possible, ensure that `decode` returns an `ndarray`. Many codecs
already do this except for some unusual ones that work with Python
objects or text data. In the cases where the data is known to be binary,
try to return `ndarray`s. This makes it pretty easy for end users of
this data to do whatever they would like with it.