Ensure Pickle's decode gets a contiguous buffer
On Python 3, `pickle.loads` expects an object that implements the buffer
protocol and is contiguous. While it would already raise this error, it
helpful for us to check this first. Plus we should be able to handle
Fortran order buffers should that be needed. On Python 2, the more
strict constraint of a `bytes` object is required, which is still
enforced.