Use (new) buffer protocol in `*MsgPack`'s `decode` (#148)
* Use (new) buffer protocol in `*MsgPack`'s `decode`

The `msgpack` implementation can leverage the (new) buffer protocol with
input data. To leverage this, `msgpack` requires that the data is
C-contiguous. To take advantage of this and avoid a copy with our data,
use `ensure_contiguous_ndarray` to take view onto the `buf` provided and
flatten it. Then pass the resulting `ndarray` to `msgpack.unpackb`,
which in turn will use the buffer protocol to perform the unpacking
without copying.

* release notes [ci skip]