Use buffer `StringIO` with `GzipFile` on Python 2
In `GZip`'s `decode` method, we knew that `BytesIO` was always copying
the data. So we merely ensured the data was in a form that made it easy
to copy (i.e. a contiguous buffer). However we have since learned we can
do better as the Python 2 implementation of `StringIO` merely takes the
buffer without copying it. So use `StringIO` on Python 2 to avoid a copy
here. As this support doesn't extend to Python 3, continue using
`BytesIO` in that case. Should avoid copies on Python 2 in the general
case when using `GZip`'s `decode`; thus, improving performance there.

ref: http://www.hydrogen18.com/blog/unpickling-buffers.html

ref: https://github.com/python/cpython/blob/2.7/Modules/cStringIO.c#L716
ref: https://github.com/python/cpython/blob/2.7/Modules/cStringIO.c#L681
ref: https://github.com/python/cpython/blob/2.7/Modules/cStringIO.c#L160