Use (new) buffer protocol in `*JSON`'s `decode` (#151)
* Use codecs.decode in JSON.decode to avoid a copy

In Python 2/3, `codecs.decode` performs decoding analogous to
`bytes.decode` except that it is not limited to `bytes`. In fact, it can
take anything that supports the (new) buffer protocol. Thus we can avoid
copying the data and instead only take an `ndarray` view onto it before
passing the buffer to `codecs.decode`. Should improve performance a bit
by avoiding this copy.

* Note JSON (new) buffer protocol support