Create MemoryViewIO class
In some cases it is very valuable to present a chunk of memory with the
file interface so as to use with functions and classes that expect a
file object. Ideally `BytesIO` would be sufficient for these needs as it
provides an in-memory file that is builtin to the standard library.
However `BytesIO` copies objects instead of preserving their underlying
buffers. When the chunks of data are large, this becomes quite unwieldy
as it uses a lot of memory and is quite slow.

To address this issue, this creates a `MemoryViewIO` class. It takes a
single buffer as input and coerces it into a contiguous memory segment
and casts it as `uint8`. Operations are provided to allow moving through
the memory as one would with a file and reading pieces of it into
`bytes`. No write support is provided currently as it is not needed.
Though this is sufficient for our needs when we want to provide a memory
segment to another function that expects a file object.