Handle datetime/timedelta case after object case
As the object check and datetime/timedelta check are exclusive and the
object check follows by raising an exception, go ahead and swap the
order of these two checks to have the object check go first. That way,
we are not spending time running the object check after we have already
determined it is datetime or timedelta and casted it to a non-object
type array. Instead we merely have verified it is a non-object array if
we got to the datetime/timedelta check, which makes it being datetime or
timedelta a possibility, but not a certainty.