Add a `subok` to the `ensure_*` functions
Provides a `subok` flag to the `ensure_*` functions much like the one
NumPy frequently uses. It's purpose is to either accept subclassed types
and work with them or not accept them and force them into the exact type
(e.g. an `ndarray` or a `bytes` object). This is useful in cases where
subclassed objects would otherwise cause problems.