Force binary data to `bytes` in `ensure_text`
To handle array types in `ensure_text` that do not support the Python
buffer protocol, it can be handy to coerce them to `bytes` using the
`tobytes` method. This way we know we have a `bytes` object that can be
decoded.

Note: This does incur a copy for some types that may not need this.