Only use `ensure_bytes` when `memoryview` fails
Instead of coercing all data to `bytes`, only perform this coercion when
the data does not support the Python buffer protocol. IOW when calling
`memoryview` on it would raise a `TypeError`.

Note: This avoids a copy when the data supports the Python buffer
protocol. Otherwise the copy is needed as `codecs.decode` will not know
how to work with it, but it can work with `bytes`.