Trim wheel builds (#320)
* Skip 32-bit wheel builds

Have not been doing testing for 32-bit or supported it for some time. So
drop 32-bit builds from wheels.

* Drop musl builds

This is a bit of an edge case that we are not otherwise testing on. So
skip these builds to lighten the build matrix.

* Drop s390x builds

This is also a bit of an edge case architecture wise. Also we are not
doing any other testing there. So drop it.