Drop `stdint_compat.h`
Everything that was relevant here has been moved to Cython. Also Cython
has its own `stdint` wrappers to handle differences encountered. So
those are not needed.

Plus this was only needed for old Visual Studio versions, which are not
used any more. In particular this was used with Python 2.7, which was
dropped from the codebase.