Drop headers (#375)
* Ignore Cython generated source files in numcodecs

* Capitalize `F`s in hex values

* Wrap `load_le32` over multiple lines

Try to match `store_le32` and improve readability.

* Replace `x` & `y` with `i` & `j`

Given these variables are integers, the latter is the more typical
naming convention.

* Use stdint types in `store_le32`/`load_le32`

* Replace `UINT32_SIZE` with `sizeof(uint32_t)`

* Implement `store_le32` & `load_le32` in Cython

* Drop `stdint_compat.h`

Everything that was relevant here has been moved to Cython. Also Cython
has its own `stdint` wrappers to handle differences encountered. So
those are not needed.

Plus this was only needed for old Visual Studio versions, which are not
used any more. In particular this was used with Python 2.7, which was
dropped from the codebase.

* Drop numcodecs headers from `MANIFEST.in`

Since the only header file included has been dropped, there is no need
to include header files. So drop this from the `MANIFEST.in` as well.

* Ignore Cython generated header files in numcodecs