numcodecs.ndarray_like: speedup is_ndarray_like (#380)
* numcodecs.ndarray_like: speedup is_ndarray_like

`is_ndarray_like()` calls `isinstance(obj, NDArrayLike)` where
`NDArrayLike` is a runtime checkable Protocol class. For every
check the protocol attributes are generated dynamically which
can significantly contribute to the decode time in zarr arrays.

This commit uses a pragmatic solution and duplicates the
protocol attributes in the code base. This could be generated
from `typing._get_protocol_attrs(NDArrayLike)` but would have
to rely on private functionality of the typing module.

* numcodecs.ndarray_like: cache results of isinstance checks

* numcodecs.ndarray_like: only cache NDArrayLike for now

* tests: check isinstance behavior for ndarray_like

* numcodecs.ndarray_like: enable caching for DType and FlagsObj

* numcodecs.ndarray_like: fix linter warnings

* update docs/release.rst

* update docs/release.rst