Delete unused `build.cmd`
This was used to configure compilers on Windows in the past.
Particularly it was used to configure Python 2.7 Windows builds. However
as of Python 3.5 and UCRT on Windows, this is unnecessary. Further
Python 2.7 was improved by a one off Windows Python compiler installer.
In any event this can safely be dropped. It is not used in this repo and
shouldn't be needed outside of it.