Ensure `entries.update(...)` recieves a `dict`
On Python 3.12, the existing code did not produce a `dict` that could be
used to update `entries` whereas previous Python versions had. To
workaround this issue, manually create a `dict` that can be used to
update `entries` with a format consistent to what had been seen in
Python versions before 3.12.