Update optimized `Delta` codec for `bool` (#595)
* Update optimized `Delta` codec for `bool`

Previously `bool` just worked with `Delta`. However this was not
actually tested. The optimized version switched to `np.subtract` for
in-place computation, which works for other types. Though `bool` needs
special handling. Fortunately this can be done with `np.not_equal`,
which has the same behavior.

Also include a test for `bool` data to make sure this is handled
correctly going forward.

* Simplify type check

This is a bit more succinct and gets to the core point. Namely
`arr.dtype` determines this code path.

Also comparing directly to `bool` works here. It is a bit faster as well
since we need not construct an `np.dtype` object.

* Drop extraneous `0`s