Wrap `_PyBytes_Resize` for improved usability
The `_PyBytes_Resize` function is helpful for resizing a `bytes` object
after it is allocated. When the underlying `bytes` object only has one
reference to it, the function can potentially use realloc to shrink or
grow the allocation in-place.

While the function signature of `_PyBytes_Resize` makes sense, it is a
little unwieldly when used directly in Cython. To smooth this out a
bit, use a macro to wrap calls to `_PyBytes_Resize`. This allows us to
work with `PyObject*`s, which Cython handles well, instead of
`PyObject**`s, which Cython handles awkwardly.

The end result is a function from Cython's perspective, which is easy to
use, and one under-the-hood that simply massages our input arguments
into something `_PyBytes_Resize` expects.