Store normalized values in `VLenBytes.encode`
During `VLenBytes.encode`, it normalizes some values. However it was not
actually keeping the normalized values. So when it comes time to get
properties from these items, it assumes they are of the right type.
However it is still grabbing the original unnormalized values. So fix
this by storing the normalized values for process.