Use Cython's tight binding in `vlen`
Instead of making so many calls to C Python APIs, rely on the Cython
types of variables and Cython's tight binding to pick the right function
to apply in each instance.

This makes the code more readable to the typical Python developer. Also
it makes it less likely some issue would sneak in like the one
encountered in this bug report.