Use `const char*` with binary encoded values
When encoding values, there is no need to modify the encoded data when
writing it out. So mark the pointers used to reference the encoded data
as `const`. While there is nothing happening here that should cause
issues, this will help further safeguard developers making changes here
and clarify the intent.